PS-5024: Prevented ICP for keys that cannot be decoded
https://jira.percona.com/browse/PS-5024

The problem was that original field value of char/varchar is not decoded for most of character sets / collations. Other types have implemented 2-way mapping functions (value <-> key) (Rdb_field_packing::setup()). utf8mb4 strings are packed by pack_with_make_sort_key() function which uses strnxfrm of MY_COLLATION_HANDLER (effectively ctype-uca.cc my_strnxfrm_uca_900()). There is no function that works in the opposite way (this is by design for all collations). This causes unpack function to be not specified in Field structure, so it is unable to decode the key and use in ICP.