PS-9753: Investigate Optimizer patches from Enhanced MySQL
https://perconadev.atlassian.net/browse/PS-9753

This commit integrates a patch included in commit 813b18c3 of
https://github.com/enhancedformysql/mysql-8.0.40
related to the optimizer patch.

Upstream commit 9a13c1c6 refactored test_quick_select() function
removing QEP_TAB dependency from it.
Now prev_tables and read_tables are passed as arguments.
get_quick_record_count() path passes nullptrs for both.
If the optimizer detected the const table in the query this information
is not passed to test_quick_select() and may result in non optimal range
scan boundaries.

This patch passes const_tables to test_quick_select(), so the behavior
for get_quick_record_count() is the same as it was before
the above-mentioned refactoring.